// src/datamodels/generatedvalue_model.tsx
export interface GeneratedDataModel {
  id: number;
  stckname: string;
  curr_month_vol?: number;
  curr_month_high?: number;
  curr_month_low?: number;
  curr_month_close?: number;
  curr_month_open?: number;
  prev_month_high?: number;
  prev_month_low?: number;
  prev_month_close?: number;
  prev_month_open?: number;
  curr_week_vol?: number;
  curr_week_high?: number;
  curr_week_low?: number;
  curr_week_close?: number;
  curr_week_open?: number;
  curr_week_rsi14?: number;
  curr_week_ema20?: number;
  curr_week_sma20?: number;
  curr_week_ema50?: number;
  curr_week_sma50?: number;
  prev_week_vol?: number;
  prev_week_high?: number;
  prev_week_low?: number;
  prev_week_close?: number;
  prev_week_open?: number;
  open: number;
  close: number;
  high: number;
  low: number;
  vol: number;
  pcnt: number;
  ema10?: number;
  ema20?: number;
  ema50?: number;
  ema100?: number;
  ema150?: number;
  ema200?: number;
  sma10?: number;
  sma20?: number;
  sma50?: number;
  sma100?: number;
  sma150?: number;
  sma200?: number;
  rsi14?: number;
  prev_day_high?: number;
  prev_day2_high?: number;
  prev_day3_high?: number;
  prev_day4_high?: number;
  prev_day5_high?: number;
  prev_day6_high?: number;
  prev_day7_high?: number;
  prev_day8_high?: number;
  prev_day9_high?: number;
  prev_day_close?: number;
  prev_day2_close?: number;
  prev_day3_close?: number;
  prev_day4_close?: number;
  prev_day5_close?: number;
  prev_day6_close?: number;
  prev_day7_close?: number;
  prev_day8_close?: number;
  prev_day9_close?: number;
  prev_day_open?: number;
  prev_day2_open?: number;
  prev_day3_open?: number;
  prev_day4_open?: number;
  prev_day5_open?: number;
  prev_day6_open?: number;
  prev_day7_open?: number;
  prev_day8_open?: number;
  prev_day9_open?: number;
  prev_day_low?: number;
  prev_day2_low?: number;
  prev_day3_low?: number;
  prev_day4_low?: number;
  prev_day5_low?: number;
  prev_day6_low?: number;
  prev_day7_low?: number;
  prev_day8_low?: number;
  prev_day9_low?: number;
  prev_day_vol?: number;
  prev_day2_vol?: number;
  prev_day3_vol?: number;
  prev_day4_vol?: number;
  prev_day5_vol?: number;
  max_250_high?: number;
  min_250_low?: number;
  prev_max_250_high?: number;
  prev_min_250_low?: number;
  prev_day_min_10_high?: number;
  prev_day2_max_20_high?: number;
  prev_day2_min_20_low?: number;
  week_max_52_high?: number;
  week_min_52_low?: number;
  curr_day_max5_close?: number;
  prev_day6_max120_close?: number;
  curr_day_sma5_volume?: number;
  sec?: string;
}